pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptstrike-cli"
version = "0.1.0"
description = "CLI tool for red-teaming LLM systems against prompt injection attacks."
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "httpx>=0.27.0"
]

[project.scripts]
promptstrike = "promptstrike.cli.main:app"

[tool.setuptools.packages.find]
include = ["promptstrike*"]
